Move listenToCommentFields to `on-comment-field-edit`
